add weight to LR_WEIGHTED_AVERAGE control in shell
 - left TC is light, right TC is heavy
 - change weight to be integer from 0 .. 100
 - add weight argument to reflow command in the shell

+++ src/sensor.h
@@ -17,6 +17,6 @@ void Sensor_DoConversion(void);
 
 float Sensor_GetTemp(TempSensor_t sensor);
 uint8_t Sensor_IsValid(TempSensor_t sensor);
-void Sensor_SetWeight(float w);
+void Sensor_SetWeight(int w);
 
 #endif /* SENSORS_H_ */
